feat: checking other costs in use

+++ src/other-costs/other-costs.service.js
@@ -72,6 +72,9 @@ class OtherCostsService {
         if (!otherCosts) {
             throw new NotFoundError("OtherCost not found")
         }
+        if (otherCosts.used) {
+            throw new ConflictError("OtherCost is in use")
+        }
         return this.otherCostsRepository.delete(uuid);
     }
 }
